+++ fiches_urgence/main.py
@@ -1,4 +1,4 @@
-from backend import create_app
+from fiches_urgence import create_app
 
 if __name__ == "__main__":
     app = create_app()
